feat: viewType을 store로 분리 #363
store에 useViewTypeStore를 만들어서 viewType을 따로 관리하도록 수정
