+++ frontend/src/components/MyDiary/ViewType.tsx
@@ -1,13 +1,21 @@
+import { useNavigate } from 'react-router-dom';
 import { viewTypes } from '@type/pages/MyDiary';
+import useViewTypeStore from '@store/useViewTypeStore';
+import { DIARY_VIEW_TYPE_LIST, PAGE_URL } from '@util/constants';
 
-import { DIARY_VIEW_TYPE_LIST } from '@util/constants';
+const ViewType = () => {
+  const { viewType, setViewType } = useViewTypeStore();
+  const navigate = useNavigate();
 
-interface ViewTypeProp {
-  viewType: viewTypes;
-  handleViewTypeChange: (type: viewTypes) => void;
-}
+  const handleViewTypeChange = (type: viewTypes) => {
+    setViewType(type);
+    navigate(`${PAGE_URL.MY_DIARY}`, {
+      state: {
+        viewType: type,
+      },
+    });
+  };
 
-const ViewType = ({ viewType, handleViewTypeChange }: ViewTypeProp) => {
   return (
     <section className="flex justify-end gap-2 pr-5 sm:justify-center sm:pr-0">
       {DIARY_VIEW_TYPE_LIST.map((type, index) => (
