+++ frontend/tsconfig.json
@@ -8,8 +8,9 @@
       "@util/*": ["./src/util/*"],
       "@assets/*": ["./src/assets/*"],
       "@api/*": ["./src/api/*"],
-      "@type/*": ["src/types/*"],
-      "@hooks/*": ["src/hooks/*"]
+      "@type/*": ["./src/types/*"],
+      "@hooks/*": ["./src/hooks/*"],
+      "@store/*": ["./src/store/*"]
     },
     "target": "ES2020",
     "useDefineForClassFields": true,
